LayoutTests/fast/canvas/canvas-overloads-setStrokeColor-expected.txt

Test the behavior of CanvasRenderingContext2D.setStrokeColor() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.setStrokeColor() threw exception TypeError: Not enough arguments.
PASS ctx.setStrokeColor('red') is undefined
PASS ctx.setStrokeColor(0) is undefined
PASS ctx.setStrokeColor(0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0) threw exception TypeError: Type error.
PASS ctx.setStrokeColor(0, 0, 0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0, 0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0, 0, 0, 0) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
